Refactor fetching purchases to the hook
